requirements.txt
boto3
python-dotenv
requests
streamlit
pytz
matplotlib
pylance
seaborn